vf_photosensitivity: add blend option
This commit attempts to address feedback from this filter's users, by
introducing a new option which controls the amelioration mechanism.

The "blend" option is a factor which is multiplied by the difference
in badness (between the threshold and the currently accumulated
badness). This difference normally controls how much of the next frame
we can let through without making it exceed the badness threshold.

Setting the option to zero effectively puts the filter into a mode
where it always duplicates the last frame (which did not put
accumulated badness over the threshold) instead of attempting to blend
in new frames.  I have received reports that this mode is preferable
to users for some types of media.